fix(vendor-batch): pin RecomputeComparisonValues divergence, complete DO-NOT-TOUCH sweep, fix overflow
Review findings on 0b60ceb (merged-ceil-remainder stream):

1. UNTESTED RUNTIME CHANGE: 0b60ceb's rewrite of
   RecomputeComparisonValues' currency-equivalent apportionment (last-
   occurrence-absorbs -> Hamilton) had no per-occurrence test - the
   only test touching this path asserted the summed ComparisonValue,
   which is identical under both algorithms for its 2x qty-1 shape.
   Added MultiOccurrenceMergedVendorOffer_ValuedCurrency_
   ComparisonValueDivergesPerOccurrenceUnderOldSharingRule (two qty-3
   occurrences, currency value 10 not evenly divisible by total
   quantity 6: old algorithm gives 3/7, new gives 5/5) to close the gap.

2. Completed the class-sweep 0b60ceb started for the self-contradicting
   "DO-NOT-TOUCH" markers: it fixed one instance (VendorBatchSolver.cs
   :873) but left four survivors (VendorBatchSolver.cs class doc,
   PlanSolver.cs class doc, and two PlanSolver.cs call-site comments)
   still asserting the merged-ceil arithmetic was frozen/unchanged.
   Corrected all four to note the 2026-08-17 retirement instead.

3. Corrected docs/KNOWN-ISSUES.md's merged-ceil-remainder milestone
   record, stale since 0b60ceb landed after it was written: added the
   omitted 5th commit (0b60ceb itself) to "What changed", corrected the
   test count (was 1773/1773; actual HEAD-at-0b60ceb was already
   1774/1774), and corrected the "0 errors, pre-existing StyleCop
   warnings only, none in any touched file" claim - measured
   PlanSolver.cs 134->142, VendorBatchSolver.cs 50->54,
   PlanViewModelBuilder.cs 158->160 across the stream (+14 new
   SA1512/SA1513/SA1515 warnings, none from this commit's own diff).

4. Fixed the overflow risk in both largest-remainder apportionment
   sites: AllocateVendorNodeCosts' `step.TotalCost * quantity` and
   RecomputeComparisonValues' `totalCurrencyValue * quantity` multiply
   a long by an int with no guard, and on overflow the numerator goes
   negative, silently breaking the "shares always sum to the total"
   invariant both doc comments assert unconditionally. Widened both
   multiply/divide sites to decimal, whose range (~7.9e28) comfortably
   covers any long x int product either field's own type can hold
   (~1.98e28 max) - removes the overflow risk entirely rather than
   just documenting it as a limitation.

Validation: build 0 errors (0 new StyleCop warnings in touched files,
verified by diffing full-rebuild output before/after); full suite
1775/1775 green (1774 baseline + 1 new characterization test).

Repo Invariants Checklist:
- [x] No Blish HUD references added to tests
- [x] Tests exercise real production paths (new test is a genuine
  Solve() round trip)
- [x] No fake file I/O tests introduced
- [x] Pricing logic preserves multi-source correctness (overflow fix
  changes no value for any realistic input, only removes wrap-around)
- [x] IDs remain internal-only (not displayed)

+++ Services/PlanSolver.cs
@@ -18,10 +18,15 @@ namespace GW2CraftingHelper.Services
         // VendorBatchesEqual, ScaleCostLines) now lives in the injected
         // VendorBatchSolver collaborator (Services/VendorBatchSolver.cs)
         // instead of this class's own private statics. Pure move - the
-        // merged-ceil arithmetic itself is DO-NOT-TOUCH
-        // (m38-cleanup-plan.md #7) and is unchanged; only the call shape
-        // moved. The parameterless constructor keeps every existing
-        // `new PlanSolver()` call site unchanged.
+        // merged-ceil arithmetic itself was DO-NOT-TOUCH at the time
+        // (m38-cleanup-plan.md #7) and was unchanged by this move; only
+        // the call shape moved. That freeze was retired 2026-08-17
+        // (characterization-first proof required, see KNOWN-ISSUES) and
+        // the arithmetic has since been rewritten (largest-remainder/
+        // Hamilton apportionment - see AllocateVendorNodeCosts' own doc
+        // comment), so it is no longer unchanged. The parameterless
+        // constructor keeps every existing `new PlanSolver()` call site
+        // unchanged.
         private readonly VendorBatchSolver _vendorBatchSolver;
 
         public PlanSolver()
@@ -477,9 +482,14 @@ namespace GW2CraftingHelper.Services
             // order from vendorOccurrences - so shares always sum to
             // precisely the step total, no drift, no invented precision,
             // and no unbounded skew toward whichever occurrence happens to
-            // land last. AllocateVendorNodeCosts (VendorBatchSolver,
-            // DO-NOT-TOUCH: merged-ceil batching math) corrects TotalCost
-            // for every occurrence of a merged vendor step but has no reason
+            // land last (the multiply widens to decimal, mirroring
+            // AllocateVendorNodeCosts' own overflow fix, so this holds
+            // unconditionally - no long overflow is possible for any
+            // totalCurrencyValue/Quantity pair either field's own type can
+            // hold). AllocateVendorNodeCosts (VendorBatchSolver; merged-ceil
+            // batching math, formerly DO-NOT-TOUCH - retired 2026-08-17, see
+            // KNOWN-ISSUES) corrects TotalCost for every occurrence of a
+            // merged vendor step but has no reason
             // to know about ComparisonValue at all; done here in the
             // PlanSolver-side wrapper instead, exactly mirroring how
             // FlagUnreliableVendorComponentCosts just below already reads
@@ -550,9 +560,19 @@ namespace GW2CraftingHelper.Services
                     long allocatedCurrency = 0L;
                     for (int i = 0; i < occurrences.Count; i++)
                     {
-                        long numerator = totalCurrencyValue * occurrences[i].Quantity;
-                        currencyShares[i] = numerator / totalQuantity;
-                        currencyRemainders[i] = numerator % totalQuantity;
+                        // Review fix (overflow): mirrors AllocateVendorNodeCosts'
+                        // own fix for the identical exposure - totalCurrencyValue
+                        // (long) * quantity (int) can exceed long range, and on
+                        // wrap the numerator goes negative, silently breaking the
+                        // sum-to-totalCurrencyValue invariant below. Widened to
+                        // decimal: the product (<= ~9.2e18 * ~2.1e9 = ~1.98e28)
+                        // always fits decimal's range (~7.9e28), so no overflow
+                        // is possible for any value either operand's own type can
+                        // hold. Both operands are whole coppers, so truncating
+                        // back to long after the divide is exact.
+                        decimal numerator = (decimal)totalCurrencyValue * occurrences[i].Quantity;
+                        currencyShares[i] = (long)(numerator / totalQuantity);
+                        currencyRemainders[i] = (long)(numerator % totalQuantity);
                         allocatedCurrency += currencyShares[i];
                     }
 
@@ -610,8 +630,9 @@ namespace GW2CraftingHelper.Services
             // vendor step, but a decision's VendorItemCosts/VendorCurrencyCosts
             // (captured pre-merge, per occurrence) are never re-derived the
             // same way - see FlagUnreliableVendorComponentCosts' own doc
-            // comment. Deliberately kept OUT of VendorBatchSolver (DO-NOT-
-            // TOUCH: merged-ceil batching math) - this only READS
+            // comment. Deliberately kept OUT of VendorBatchSolver (merged-
+            // ceil batching math, formerly DO-NOT-TOUCH - retired
+            // 2026-08-17, see KNOWN-ISSUES) - this only READS
             // AllocateVendorNodeCosts' own already-public inputs/outputs
             // (vendorOccurrences, stepMap) after it returns, and writes a
             // new auxiliary flag; it changes no cost, no share, no batch
@@ -2230,7 +2251,8 @@ namespace GW2CraftingHelper.Services
         /// Read-only with respect to VendorBatchSolver: only inspects
         /// stepMap/vendorOccurrences and writes the new auxiliary flag on
         /// `memo` - never touches TotalCost, UnitCost, or any batch/ceil
-        /// arithmetic (all DO-NOT-TOUCH, computed entirely by
+        /// arithmetic (merged-ceil batching math, formerly DO-NOT-TOUCH -
+        /// retired 2026-08-17, see KNOWN-ISSUES; still computed entirely by
         /// AllocateVendorNodeCosts/FinalizeVendorBatches above).
         /// </summary>
         private static void FlagUnreliableVendorComponentCosts(
